ffi: Refactor authentication_service.rs to authentication.rs
