+++ bindings/matrix-sdk-ffi/src/lib.rs
@@ -20,7 +20,7 @@ macro_rules! unwrap_or_clone_arc_into_variant {
     };
 }
 
-mod authentication_service;
+mod authentication;
 mod chunk_iterator;
 mod client;
 mod client_builder;
